Refactor duplicate code, CreateVirtualNodeRequest in a sealed class
