+++ components/virtual-node/virtual-node-rest-service-impl/src/main/kotlin/net/corda/virtualnode/rest/factories/RequestFactory.kt
@@ -2,23 +2,15 @@ package net.corda.virtualnode.rest.factories
 
 import net.corda.data.virtualnode.VirtualNodeAsynchronousRequest
 import net.corda.libs.virtualnode.endpoints.v1.types.CreateVirtualNodeRequest
-import net.corda.libs.virtualnode.endpoints.v1.types.JsonCreateVirtualNodeRequest
 import net.corda.libs.virtualnode.endpoints.v1.types.UpdateVirtualNodeDbRequest
 import net.corda.virtualnode.HoldingIdentity
 
 internal interface RequestFactory {
     fun createHoldingIdentity(groupId: String, request: CreateVirtualNodeRequest): HoldingIdentity
 
-    fun createHoldingIdentity(groupId: String, request: JsonCreateVirtualNodeRequest): HoldingIdentity
-
     fun createVirtualNodeRequest(
         holdingIdentity: HoldingIdentity,
         request: CreateVirtualNodeRequest
-    ): VirtualNodeAsynchronousRequest
-
-    fun createVirtualNodeRequest(
-        holdingIdentity: HoldingIdentity,
-        request: JsonCreateVirtualNodeRequest
     ): VirtualNodeAsynchronousRequest
 
     fun updateVirtualNodeDbRequest(
