+++ components/virtual-node/virtual-node-rest-service-impl/src/main/kotlin/net/corda/virtualnode/rest/impl/validation/VirtualNodeValidationService.kt
@@ -2,14 +2,12 @@ package net.corda.virtualnode.rest.impl.validation
 
 import net.corda.libs.packaging.core.CpiMetadata
 import net.corda.libs.virtualnode.endpoints.v1.types.CreateVirtualNodeRequest
-import net.corda.libs.virtualnode.endpoints.v1.types.JsonCreateVirtualNodeRequest
 import net.corda.virtualnode.HoldingIdentity
 import net.corda.virtualnode.VirtualNodeInfo
 
 internal interface VirtualNodeValidationService {
     fun validateVirtualNodeDoesNotExist(holdingIdentity: HoldingIdentity)
     fun validateAndGetGroupId(request: CreateVirtualNodeRequest): String
-    fun validateAndGetGroupId(request: JsonCreateVirtualNodeRequest): String
     fun validateAndGetVirtualNode(virtualNodeShortId: String): VirtualNodeInfo
     fun validateAndGetCpiByChecksum(cpiFileChecksum: String): CpiMetadata
     fun validateCpiUpgradePrerequisites(currentCpi: CpiMetadata, upgradeCpi: CpiMetadata)
